refactor(admin): TV links on dashboard; kiosk-only wall view
Embed TV link create/list/revoke on main analytics dashboard. Remove
standalone /dashboard/tv and /dashboard/tv-links. Wall UI is only via
/tv/view/<token>. Also harden TTL validation, list limit, auth path,
snapshot calendar rates, clear board on poll error.
